Upgrade to Ruby 2.3.0
In this commit, we are upgrading Ruby from 2.1.9 to 2.3.0.

We ae also upgrading the following dependencies that no longer work
with Ruby 2.3.0:
- rails: 3.1.0 => 3.2.22
- compass: 0.12.alpha.0 => 1.0.3
- sass-rails: 3.1.0 => 3.2.0

We had to add test-unit to the dependency list since should-matchers
requires it to work.

We are also removing calls to compass that were failing.

The system works, but the styles are a bit off.